README.md
Financial Data Analysis Dashboard

This repository contains a Power BI dashboard that analyzes financial data across different countries, products, and time dimensions. The dashboard provides insights into sales, profits, and units sold.

## Power BI Dashboard

The Power BI dashboard file (PBIX) can be found in the [dashboard](/dashboard) folder. It showcases visualizations and interactive reports for analyzing financial performance.

## Getting Started

To view and interact with the dashboard, follow these steps:

1. Download the PBIX file from the [Releases](/releases) section or directly from the [dashboard](/dashboard) folder.
2. Open the PBIX file in Power BI Desktop.
3. Connect the dashboard to the financial dataset provided.
4. Explore the different pages, filters, and visuals to gain insights into sales, profits, and units sold across countries, products, and time dimensions.


## Dataset

The dataset used in this dashboard is available here. It includes financial data for various countries, products, months, and years. The dataset file is provided in XLSX format.

## 'Images': 
Includes preview and background images of the dashboard.

Feel free to contribute to this repository by adding new analyses or improvements to the dashboard. Feedback and suggestions are always welcome!